app/requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-filters==0.2.1
djangorestframework==3.6.4
mysqlclient==1.3.12
pytz==2017.2
